requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pandas==2.2.3
openpyxl==3.1.5
aiohttp==3.10.5
python-multipart==0.0.12
pydantic==2.9.2
python-dotenv==1.0.1